+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 * Added item tooltip display to compacting drawers and a proper capability, closes #48
 * Fixed Simple Compacting Drawers being breakable in creative when interacting with them, closes #187
 * Prevent infinite loop when drawers check for their parent stuff, closes #73
+* Now having the configuration tool in the offhand it will toggle the selected action when placing a drawer, closes #156